refactor: student_view_router를 classroom에서 progress 도메인으로 이동
classroom/student_view_router.py는 이름과 위치 모두 잘못되어 있었다.
해당 라우터는 교사가 학생의 학습 기록과 약점 프로필을 조회하는 API로,
"반 관리"가 아닌 "학습 기록 조회(progress)" 도메인의 책임이다.

변경 내용:
- classroom/student_view_router.py → progress/teacher_router.py 이동
- WeakConceptResponse, StudentWeaknessProfileResponse를 agent/schemas.py에서
  progress/schemas.py로 이동 (약점 프로필은 progress 도메인 개념)
- agent/schemas.py는 progress/schemas.py에서 re-export (기존 코드 호환 유지)

의존성 방향 변화:
  before: classroom → progress, classroom → agent (위반)
  after:  progress → classroom (접근 권한 확인용), agent → progress (정방향)
