+++ app/domains/progress/teacher_router.py
@@ -1,39 +1,41 @@
-"""교사가 특정 학생의 학습 상태를 조회하는 Classroom 라우터.
+"""교사가 담당 학생의 학습 상태를 조회하는 라우터.
 
 경로 prefix: /teacher/students
 주 사용자: 교사, 개발자
 
-teacher_router.py와 분리한 이유:
-- teacher_router.py는 반 자체를 관리한다. 예: 반 생성, 반 목록, 학생 추가/삭제.
-- 이 파일은 교사가 담당 학생 1명의 프로필/학습 기록을 보는 read-only 조회 API를 담는다.
-- URL도 /teacher/classes가 아니라 /teacher/students/{user_id} 중심이다.
+progress 도메인에 위치한 이유:
+- 학생의 약점 프로필과 학습 기록 조회는 progress 도메인의 책임이다.
+- 교사 권한 확인은 classroom 서비스에 위임하고, 데이터는 progress 서비스에서 가져온다.
 """
 from fastapi import APIRouter, Depends, HTTPException, Query, status
 
-from app.domains.agent.schemas import AgentProfileResponse, WeakConceptResponse
 from app.domains.auth.dependencies import get_current_teacher
 from app.domains.auth.models import User
 from app.domains.classroom.dependencies import get_classroom_service
 from app.domains.classroom.service import ClassroomService
 from app.domains.progress.dependencies import get_learning_record_service
-from app.domains.progress.schemas import LearningRecordResponse, LearningRecordsResponse
+from app.domains.progress.schemas import (
+    LearningRecordResponse,
+    LearningRecordsResponse,
+    StudentWeaknessProfileResponse,
+    WeakConceptResponse,
+)
 from app.domains.progress.service import LearningRecordService
 
-# 교사 관점의 학생 상세 조회 API.
 router = APIRouter(prefix="/teacher/students", tags=["teacher"])
 
 
-@router.get("/{user_id}/profile", response_model=AgentProfileResponse)
+@router.get("/{user_id}/profile", response_model=StudentWeaknessProfileResponse)
 def get_student_profile(
     user_id: str,
     current_user: User = Depends(get_current_teacher),
     classroom_service: ClassroomService = Depends(get_classroom_service),
     learning_record_service: LearningRecordService = Depends(get_learning_record_service),
-) -> AgentProfileResponse:
+) -> StudentWeaknessProfileResponse:
     """담당 학생의 약점 개념 프로필을 조회한다."""
     _ensure_student_access(classroom_service, current_user, user_id)
     profile = learning_record_service.get_weakness_profile(user_id)
-    return AgentProfileResponse(
+    return StudentWeaknessProfileResponse(
         user_id=profile.user_id,
         weak_concepts=[
             WeakConceptResponse(**weak_concept.model_dump())
@@ -73,7 +75,7 @@ def _ensure_student_access(
     current_user: User,
     student_id: str,
 ) -> None:
-    """교사가 담당하지 않는 학생의 학습 데이터에 접근하지 못하게 막는다."""
+    """교사가 담당하지 않는 학생의 데이터에 접근하지 못하게 막는다."""
     if not classroom_service.can_access_student(current_user, student_id):
         raise HTTPException(
             status_code=status.HTTP_403_FORBIDDEN,
